requirements.txt
Flask
Flask-Cors
Flask-Login
Jinja2
Pillow
requests
Werkzeug
